requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
diff-match-patch==20121119
Django==2.1.2
django-cors-middleware==1.3.1
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-import-export==1.1.0
django-rest-swagger==2.2.0
django-reversion==3.0.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.8
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
pytz==2018.5
PyYAML==3.13
requests==2.19.1
simplejson==3.16.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
xadmin==2.0.1
xlrd==1.1.0
xlwt==1.3.0
